+++ app/assets/scripts/utils/constants.js
@@ -1,5 +1,13 @@
+'use strict';
+
 export const taskStatusMatrix = {
   open: 'Open',
   inprogress: 'In progress',
   completed: 'Completed'
 };
+
+export const taskStatusStyles = [
+  {name: 'open', color: '#00F'},
+  {name: 'completed', color: '#0F0'},
+  {name: 'closed', color: '#F00'}
+];
